+++ packages/openapi/.npmignore
@@ -1,0 +1,5 @@
+src/
+test/
+
+jest.config.ts
+tsconfig.json
