+++ packages/openapi/README.md
@@ -1,0 +1,110 @@
+[![snyk](https://snyk.io/test/github/teunmooij/payload/badge.svg)](https://snyk.io/test/github/teunmooij/payload)
+[![npm version](https://badge.fury.io/js/payload-swagger.svg)](https://badge.fury.io/js/payload-swagger)
+
+# payload-swagger
+
+Swagger plugin for payload cms:
+
+- generate openAPI 3 documentation from your payload config
+- includes Swagger UI
+
+Version 0.x releases should be considered `beta`. Semantic version is not guaranteed until we reach version 1.
+
+## Installation
+
+With yarn:
+
+```shell
+yarn add payload-swagger
+```
+
+With npm:
+
+```shell
+npm install payload-swagger
+```
+
+## Usage
+
+`payload-swagger` can be used as plugin or just to generate the openapi document.
+
+### As plugin
+
+```typescript
+import { buildConfig } from 'payload/config';
+import path from 'path';
+import { swagger } from 'payload-swagger';
+
+export default buildConfig({
+  plugins: [
+    swagger({
+      /* see options section */
+    }),
+  ],
+  // The rest of your config goes here
+});
+```
+
+### Just create the openapi document
+
+`payload-swagger` creates openapi documents, version 3.0, as javascript object. If you need a yaml file, please use a library like `yaml` to convert it.
+
+```typescript
+import { createDocument } from 'payload-swagger';
+import payloadConfig from '...';
+
+const openapiDocument = createDocument(payloadConfig, {
+  /* see options section */
+});
+```
+
+## Extend the openAPI document
+
+The openapi document generated by `payload-swagger` includes information from the following external sources:
+
+- `package.json`: `payload-swagger` uses the information in the fields shown below. In `openapi` a partial openapi document can be included;
+
+```json
+{
+  "name": "my payload cms",
+  "version": "1.2.3",
+  "description": "My awesome cms",
+  "license": "MIT",
+  "openapi": {}
+}
+```
+
+- `.openapi`: optional file with partial openapi document in JSON format.
+
+The data from all these source is recursively merged into the final document.
+
+## Options
+
+`payload-swagger` has the following options:
+
+```typescript
+interface Options {
+  /**
+   * By default the access functions on all collections in the config are called to determine
+   * the access level of the operations.
+   * Provide an array of collection slugs to disable this for the given collections,
+   * or `true` to disable for all.
+   */
+  disableAccessAnalysis?: boolean | string[];
+}
+```
+
+## Version history
+
+### 0.2.x
+
+- dropped server extension in favor of plugin
+
+### 0.1.x
+
+- Added custom endpoints
+- Refactored `initSwagger` parameters
+
+### 0.0.x
+
+- Initial version
