+++ packages/openapi/src/base-config/error-response.ts
