+++ packages/openapi/src/index.ts
@@ -1,0 +1,6 @@
+import { createDocument } from './open-api';
+
+export type { Options } from './types';
+
+export { createDocument };
+export default createDocument;
