+++ packages/openapi/src/payload-config/access-path.ts
