+++ packages/openapi/src/payload-config/auth-paths.ts
