+++ packages/openapi/src/payload-config/collection-paths.ts
