+++ packages/openapi/src/payload-config/custom-paths.ts
