+++ packages/openapi/src/payload-config/global-paths.ts
