+++ packages/openapi/src/payload-config/index.ts
