+++ packages/openapi/src/schemas/basic-components.ts
