+++ packages/openapi/src/schemas/index.ts
