+++ packages/openapi/src/schemas/me.ts
