+++ packages/openapi/src/schemas/paginated-documents.ts
