+++ packages/openapi/src/schemas/upsert-confirmation.ts
