+++ packages/openapi/src/utils/entity-to-json-schema.ts
