+++ packages/openapi/src/utils/get-description.ts
