+++ packages/openapi/src/utils/index.ts
