+++ packages/openapi/src/utils/types.ts
