+++ packages/openapi/test/index.spec.ts
@@ -1,0 +1,5 @@
+describe('tests', () => {
+  it('test', () => {
+    expect(true).toBe(true);
+  });
+});
