+++ packages/swagger/src/index.ts
@@ -2,7 +2,5 @@ import { swagger } from './plugin';
 
 export type { Options } from './types';
 
-export { createDocument } from './open-api';
-
 export { swagger };
 export default swagger;
