+++ packages/swagger/src/utils/dummy-module.ts
